# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Device:
    properties:
      batteryLevel:
        example: 50
        type: integer
      batteryStatus:
        example: charging
        type: string
      name:
        example: iphone 99xs
        type: string
      time:
        example: "2006-01-02 15:04:05"
        type: string
    required:
    - batteryLevel
    - batteryStatus
    - name
    - time
    type: object
  models.DeviceDetail:
    properties:
      name:
        type: string
      os:
        type: string
      type:
        type: string
    type: object
  models.JSONfailResult:
    properties:
      code:
        example: 400
        type: integer
      data: {}
      message:
        example: error
        type: string
    type: object
  models.JSONsuccessResult:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: success
        type: string
    type: object
host: localhost
info:
  contact:
    email: aglide100@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple Battery level checker server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Battery level checker API
  version: 0.0.1
paths:
  /battery/:
    get:
      consumes:
      - application/json
      description: Get devices's battery
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONsuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Device'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONfailResult'
      summary: Get All Device's Battery Level
      tags:
      - Battery
  /battery/{deviceID}:
    get:
      consumes:
      - application/json
      description: Get devices's battery
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.JSONsuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/models.Device'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONfailResult'
      summary: Get Battery Level
      tags:
      - Battery
    put:
      consumes:
      - application/json
      description: Update devices's battery
      parameters:
      - description: Battery input form
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/models.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JSONsuccessResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONfailResult'
      summary: Update Battery Level
      tags:
      - Battery
  /device:
    post:
      consumes:
      - application/json
      description: Register New Device
      parameters:
      - description: add device
        in: body
        name: deviceInfo
        required: true
        schema:
          $ref: '#/definitions/models.DeviceDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JSONsuccessResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONfailResult'
      summary: Register New Device
      tags:
      - Device
  /device/{deviceID}:
    delete:
      consumes:
      - application/json
      description: Delete Device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JSONsuccessResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.JSONfailResult'
      summary: Delete Device
      tags:
      - Device
swagger: "2.0"